refactor: bring NavbarIcon component in to components/layout folder, remove not needed files/folders
