+++ frontend/src/components/layout/NavbarLink/NavbarLink.styles.ts
@@ -1,4 +1,5 @@
 import { NavLink } from 'react-router-dom';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import styled from 'styled-components';
 
 const LinkContainer = styled(NavLink)`
@@ -10,12 +11,28 @@ const LinkContainer = styled(NavLink)`
     text-decoration: underline;
     cursor: pointer;
   }
+  
+  & svg {
+    
+  }
 
   @media screen and (width < ${({ theme }) => theme.breakpoint.large}) {
     font-size: ${({ theme }) => theme.fontSize.large};
   }
 `;
 
+type ComponentPropsT = {
+  $isActive: boolean;
+}
+
+const NavbarIcon = styled(FontAwesomeIcon)<ComponentPropsT>`
+  display: inline-block;
+  margin: 0 1.5rem 0 0;
+  color: ${({ theme, $isActive }) => $isActive ? theme.color.tertiaryLight : theme.color.primaryDark};
+  font-size: ${({ theme }) => theme.fontSize.large};
+`;
+
 export {
   LinkContainer,
+  NavbarIcon,
 };
