+++ frontend/src/components/layout/PublicLayout/PublicLayout.component.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
-import { NavbarLink } from '@components/shared/navigation';
 import { GeneralIcon } from '@components/shared/icon-styles';
 import BaseNavbarStyle from '../BaseNavbarStyle';
+import NavbarLink from '../NavbarLink';
 import Footer from '../Footer';
 import { iconLibraryConfig } from '@configuration';
