--- frontend/src/components/shared/icon-styles/NavbarIcon/index.ts
@@ -1,3 +1,0 @@
-import NavbarIcon from './NavbarIcon.styles.ts';
-
-export default NavbarIcon;
