+++ frontend/src/components/shared/icon-styles/index.ts
@@ -1,9 +1,7 @@
 import LabelIcon from './LabelIcon';
 import GeneralIcon from './GeneralIcon';
-import NavbarIcon from './NavbarIcon';
 
 export {
   LabelIcon,
   GeneralIcon,
-  NavbarIcon,
 };
