--- frontend/src/components/shared/navigation/index.ts
@@ -1,5 +1,0 @@
-import NavbarLink from './NavbarLink';
-
-export {
-  NavbarLink,
-};
